requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone
pypdf
tqdm
pydantic